ai-room-service/requirements.txt
fastapi
uvicorn
onnxruntime
numpy
pillow
requests
